feat(backend): Payment processing, Reviews & Wishlist modules
This commit adds comprehensive payment processing, product reviews, and wishlist functionality to the e-commerce platform:

**Payment Enhancements:**
- Added PaymentAuthorizedSubscriber for automatic order confirmation after successful payment
- Enhanced Payment domain model with authorization tracking
- Improved StripePaymentController with better error handling and logging
- Updated PaymentAuthorized event with orderId context

**New Review Module:**
- Complete CQRS implementation (Commands: Delete, Update)
- Query handlers for review stats, rating distribution, and review retrieval
- REST API endpoints for review management
- Review presentation layer with DTOs

**New Wishlist Module:**
- Full bounded context implementation
- Domain models and repository interfaces
- Application layer commands and queries
- Infrastructure persistence layer

**Infrastructure Updates:**
- Added CorsPreflightListener for better CORS handling in development
- Updated CORS configuration for storefront integration
- Enhanced Doctrine configuration with custom types
- Updated security configuration for new endpoints

**Event-Driven Features:**
- OrderPlacedCartClearingSubscriber automatically clears cart after order placement
- Enhanced DomainEventAuditSubscriber for better audit logging

**Database Migrations:**
- Version20251105064800: Payment tables
- Version20251105070000: Review tables
- Version20251106083829: Wishlist tables

+++ src/Order/Infrastructure/Persistence/Doctrine/Entity/OrderEntity.php
@@ -51,8 +51,8 @@ class OrderEntity
     #[ORM\Column(type: 'json', nullable: true, name: 'applied_promotions')]
     private ?array $appliedPromotions = null;
 
-    #[ORM\Column(type: 'float', nullable: true, name: 'discount_amount')]
-    private ?float $discountAmount = null;
+    #[ORM\Column(type: 'integer', nullable: true, name: 'discount_amount')]
+    private ?int $discountAmount = null;
 
     #[ORM\Column(type: 'string', length: 3, nullable: true, name: 'discount_currency')]
     private ?string $discountCurrency = null;
@@ -60,8 +60,8 @@ class OrderEntity
     #[ORM\Column(type: 'string', length: 20, nullable: true, name: 'coupon_code')]
     private ?string $couponCode = null;
 
-    #[ORM\Column(type: 'float', nullable: true, name: 'tax_amount')]
-    private ?float $taxAmount = null;
+    #[ORM\Column(type: 'integer', nullable: true, name: 'tax_amount')]
+    private ?int $taxAmount = null;
 
     #[ORM\Column(type: 'string', length: 3, nullable: true, name: 'tax_currency')]
     private ?string $taxCurrency = null;
